workers/requirements.txt
fastapi
uvicorn
boto3
firebase-admin
psycopg2-binary
python-dotenv
